[ui] Make KeyFrameAnimation update animations itself
