+++ src/modm/ui/animation/key_frame_impl.hpp
@@ -15,23 +15,23 @@
 
 // ----------------------------------------------------------------------------
 template< typename T, class... Args >
-modm::ui::KeyFrameAnimationBase<T, Args...>::KeyFrameAnimationBase(Animation<Args>&... animator)
-:	KeyFrameAnimationBase(nullptr,0, animator...)
+modm::ui::KeyFrameAnimationBase<T, Args...>::KeyFrameAnimationBase(Animation<Args>&... animators)
+:	KeyFrameAnimationBase(nullptr,0, animators...)
 {
 }
 
 template< typename T, class... Args >
 template< uint16_t N >
-modm::ui::KeyFrameAnimationBase<T, Args...>::KeyFrameAnimationBase(const KeyFrame<T, size> (&frames)[N], Animation<Args>&... animator)
-:	KeyFrameAnimationBase(frames, N, animator...)
+modm::ui::KeyFrameAnimationBase<T, Args...>::KeyFrameAnimationBase(const KeyFrame<T, size> (&frames)[N], Animation<Args>&... animators)
+:	KeyFrameAnimationBase(frames, N, animators...)
 {
 }
 
 template< typename T, class... Args >
 modm::ui::KeyFrameAnimationBase<T, Args...>::KeyFrameAnimationBase(
 		const KeyFrame<T, size> *frames, uint16_t length,
-		Animation<Args>&... animator)
-:	animator{(&animator)...}, frames(frames), length(length),
+		Animation<Args>&... animators)
+:	animators{(&animators)...}, frames(frames), length(length),
 	currentFrame(0), repeat(0), mode(modeCycleMask)
 {
 }
@@ -63,7 +63,7 @@ modm::ui::KeyFrameAnimationBase<T, Args...>::stop()
 		for (uint_fast8_t ii = 0; ii < size; ii++)
 		{
 			// it we are running reversed, the "last" frame is the first one!
-			animator[ii]->setValue(frames[(mode & reversedMask) ? 0 : length-1].value[ii]);
+			animators[ii]->setValue(frames[(mode & reversedMask) ? 0 : length-1].value[ii]);
 		}
 	}
 	cancel();
@@ -76,9 +76,9 @@ modm::ui::KeyFrameAnimationBase<T, Args...>::cancel()
 	// disable
 	currentFrame = 0;
 	mode &= ~reversedMask;
-	for (uint_fast8_t ii = 0; ii < size; ii++)
+	for (auto *animator : animators)
 	{
-		animator[ii]->stop();
+		animator->stop();
 	}
 }
 
@@ -97,12 +97,12 @@ modm::ui::KeyFrameAnimationBase<T, Args...>::update()
 	if (currentFrame > 0)
 	{
 		// wait until the key frame finished
-		if (!animator[0]->isAnimating())
+		if (!animators[0]->isAnimating())
 		{
 			// set the next keyframe
 			for (uint_fast8_t ii = 0; ii < size; ii++)
 			{
-				animator[ii]->animateTo(
+				animators[ii]->animateTo(
 						frames[currentFrame-1].value[ii],
 						frames[currentFrame-1].length);
 			}
@@ -154,6 +154,8 @@ modm::ui::KeyFrameAnimationBase<T, Args...>::update()
 			}
 		}
 	}
+	// Update all animators
+	for (auto *animator : animators) animator->update();
 }
 
 template< typename T, class... Args >
